added CartTimer Component
