+++ src/features/Cart/index.js
@@ -1,56 +1,19 @@
-import { useEffect, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { Main } from "../../common/Main/Main";
-import {
-  removeFromCart,
-  selectCart,
-  removeAllProductFromCart,
-} from "./cartSlice";
+import { removeFromCart, selectCart } from "./cartSlice";
 import { CartListWrapper } from "../../common/Tile/styled";
 import { CartsTile } from "../../common/Tile";
 import { RemoveButton } from "../../common/Button/styled";
 import { CartText } from "./styled";
+import { CartTimer } from "./CartTimer";
 
 export const Cart = () => {
   const dispatch = useDispatch();
   const cartProducts = useSelector(selectCart);
-  const [timeLeft, setTimeLeft] = useState(() => {
-    const storeData = JSON.parse(localStorage.getItem("cart"));
-    if (storeData?.expirationTime) {
-      const remainingTime = Math.floor(
-        (storeData.expirationTime - Date.now()) / 1000
-      );
-      return remainingTime > 0 ? remainingTime : 0;
-    }
-    return 0;
-  });
-
-  useEffect(() => {
-    if (timeLeft <= 0) {
-      localStorage.removeItem("cart");
-      dispatch(removeAllProductFromCart());
-      return;
-    }
-
-    const interval = setInterval(() => {
-      setTimeLeft((prev) => (prev > 0 ? prev - 1 : 0));
-    }, 1000);
-
-    return () => clearInterval(interval);
-  }, [timeLeft, dispatch]);
-
-  const minutes = Math.floor(timeLeft / 60);
-  const seconds = timeLeft % 60;
 
   return (
     <Main>
-      {timeLeft > 0 ? (
-        <p>
-          Pozostały czas {minutes ? minutes : null} minut {seconds} sekund
-        </p>
-      ) : (
-        ""
-      )}
+      {cartProducts.length !== 0 ? <CartTimer /> : null}
       <CartListWrapper>
         {cartProducts.length === 0 ? (
           <CartText>Cart is empty!</CartText>
